+++ NEWS.md
@@ -16,6 +16,7 @@
     - Completions from packages in `library`, `require` calls automatically
       inferred and supplied even when not loaded
     - Completions for knitr options, e.g. in `opts_chunk$get()`, are now supplied
+    - Completions for dynamic symbols within `.C`, `.Call`, `.Fortran`, `.External`
 * Improvements to C/C++ editing mode:
     - Code completion
     - F2 code navigation (go to definition)